Improved "delete" (Task/User) and "update" (Task) operations

+++ src/main/java/com/business_crab/ToDoList/API/services/TaskService.java
@@ -1,6 +1,7 @@
 package com.business_crab.ToDoList.API.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -8,6 +9,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import com.business_crab.ToDoList.API.entities.Task;
 import com.business_crab.ToDoList.API.repositories.TaskRepository;
+
+import jakarta.transaction.Transactional;
 
 @Service
 public class TaskService {
@@ -25,17 +28,31 @@ public class TaskService {
         return taskRepository.save(task);
     }
 
-    public Task update(final @RequestBody Task newTask , final @PathVariable Long id) {
-        return taskRepository.findById(id).map((task) -> {
-            task.setTitle(newTask.getTitle());
-            task.setDesc(newTask.getDesc());
-            return taskRepository.save(task);
-        }).orElseGet(() -> {
-            return taskRepository.save(newTask);
-        });
+    @Transactional
+    public void update(final @PathVariable Long id ,
+                       final String title ,
+                       final String description) {
+        Optional<Task> optionalTask = taskRepository.findById(id);
+        if (optionalTask.isEmpty()) {
+            throw new IllegalStateException("The user with such id is not found");
+        }
+
+        final Task task = optionalTask.get();
+        
+        if (title != null && !title.equals(task.getTitle())) {
+            task.setTitle(title);
+        }
+
+        if (description != null && !description.equals(task.getDesc())) {
+            task.setDesc(description);
+        }
     }
     
     public void delete(final @PathVariable Long id) {
+        Optional<Task> optionalTask = taskRepository.findById(id);
+        if (optionalTask.isEmpty()) {
+            throw new IllegalStateException("Task with such id is not found");
+        }
         taskRepository.deleteById(id);
     }
 }
